add support  for general taskqueue.
taskqueue is a singletons,you may get reference to a taskqueue or create one by call spd_taskqueue_get.
this will create a singleton taskqueue ,a thread will be create to cond wait the task queue of this singleton, onece you push a task to this singleton taskqueue , the thread will pop it and call the handler function you specified when you push a task.